Large layout changes
Firstly I moved all of the image files into the imgs folder.  Please do
this to reduce clutter around our main program files.

I scaled the tab bar so that it now spans the entire width of the
screen.

Due to the tab bar functionality, I scrapped the main menu as it seemed
useless now.  I recycled the code to create an options page.

Tapping the tabs now takes you to the respective screen, so the time tab
takes you to the scheduler, gear to the options, ect...

The controltab and controltabover were named incorrectly, I renamed
them.

The tabs interface is now initiated in the controls page which is now
the first screen to appear when you start the application.